move function modules to separate files (to avoid conflict in test env)
